Separated algorithms from pub/privkey and sig declarations
