+++ ed25519.go
@@ -1,0 +1,228 @@
+package crypto
+
+import (
+	"bytes"
+	"fmt"
+
+	"golang.org/x/crypto/ripemd160"
+
+	"github.com/tendermint/ed25519"
+	"github.com/tendermint/ed25519/extra25519"
+	cmn "github.com/tendermint/go-common"
+	data "github.com/tendermint/go-data"
+	wire "github.com/tendermint/go-wire"
+)
+
+func init() {
+	privKeyMapper.RegisterImplementation(PrivKeyEd25519{}, NameEd25519, TypeEd25519)
+	pubKeyMapper.RegisterImplementation(PubKeyEd25519{}, NameEd25519, TypeEd25519)
+	sigMapper.RegisterImplementation(SignatureEd25519{}, NameEd25519, TypeEd25519)
+}
+
+//-------------------------------------
+
+var _ PrivKeyInner = PrivKeyEd25519{}
+
+// Implements PrivKey
+type PrivKeyEd25519 [64]byte
+
+func (privKey PrivKeyEd25519) AssertIsPrivKeyInner() {}
+
+func (privKey PrivKeyEd25519) Bytes() []byte {
+	return wire.BinaryBytes(PrivKey{privKey})
+}
+
+func (privKey PrivKeyEd25519) Sign(msg []byte) Signature {
+	privKeyBytes := [64]byte(privKey)
+	signatureBytes := ed25519.Sign(&privKeyBytes, msg)
+	return SignatureEd25519(*signatureBytes).Wrap()
+}
+
+func (privKey PrivKeyEd25519) PubKey() PubKey {
+	privKeyBytes := [64]byte(privKey)
+	pubBytes := *ed25519.MakePublicKey(&privKeyBytes)
+	return PubKeyEd25519(pubBytes).Wrap()
+}
+
+func (privKey PrivKeyEd25519) Equals(other PrivKey) bool {
+	if otherEd, ok := other.Unwrap().(PrivKeyEd25519); ok {
+		return bytes.Equal(privKey[:], otherEd[:])
+	} else {
+		return false
+	}
+}
+
+func (p PrivKeyEd25519) MarshalJSON() ([]byte, error) {
+	return data.Encoder.Marshal(p[:])
+}
+
+func (p *PrivKeyEd25519) UnmarshalJSON(enc []byte) error {
+	var ref []byte
+	err := data.Encoder.Unmarshal(&ref, enc)
+	copy(p[:], ref)
+	return err
+}
+
+func (privKey PrivKeyEd25519) ToCurve25519() *[32]byte {
+	keyCurve25519 := new([32]byte)
+	privKeyBytes := [64]byte(privKey)
+	extra25519.PrivateKeyToCurve25519(keyCurve25519, &privKeyBytes)
+	return keyCurve25519
+}
+
+func (privKey PrivKeyEd25519) String() string {
+	return cmn.Fmt("PrivKeyEd25519{*****}")
+}
+
+// Deterministically generates new priv-key bytes from key.
+func (privKey PrivKeyEd25519) Generate(index int) PrivKeyEd25519 {
+	newBytes := wire.BinarySha256(struct {
+		PrivKey [64]byte
+		Index   int
+	}{privKey, index})
+	var newKey [64]byte
+	copy(newKey[:], newBytes)
+	return PrivKeyEd25519(newKey)
+}
+
+func (privKey PrivKeyEd25519) Wrap() PrivKey {
+	return PrivKey{privKey}
+}
+
+func GenPrivKeyEd25519() PrivKeyEd25519 {
+	privKeyBytes := new([64]byte)
+	copy(privKeyBytes[:32], CRandBytes(32))
+	ed25519.MakePublicKey(privKeyBytes)
+	return PrivKeyEd25519(*privKeyBytes)
+}
+
+// NOTE: secret should be the output of a KDF like bcrypt,
+// if it's derived from user input.
+func GenPrivKeyEd25519FromSecret(secret []byte) PrivKeyEd25519 {
+	privKey32 := Sha256(secret) // Not Ripemd160 because we want 32 bytes.
+	privKeyBytes := new([64]byte)
+	copy(privKeyBytes[:32], privKey32)
+	ed25519.MakePublicKey(privKeyBytes)
+	return PrivKeyEd25519(*privKeyBytes)
+}
+
+//-------------------------------------
+
+var _ PubKeyInner = PubKeyEd25519{}
+
+// Implements PubKeyInner
+type PubKeyEd25519 [32]byte
+
+func (pubKey PubKeyEd25519) AssertIsPubKeyInner() {}
+
+func (pubKey PubKeyEd25519) Address() []byte {
+	w, n, err := new(bytes.Buffer), new(int), new(error)
+	wire.WriteBinary(pubKey[:], w, n, err)
+	if *err != nil {
+		cmn.PanicCrisis(*err)
+	}
+	// append type byte
+	encodedPubkey := append([]byte{TypeEd25519}, w.Bytes()...)
+	hasher := ripemd160.New()
+	hasher.Write(encodedPubkey) // does not error
+	return hasher.Sum(nil)
+}
+
+func (pubKey PubKeyEd25519) Bytes() []byte {
+	return wire.BinaryBytes(PubKey{pubKey})
+}
+
+func (pubKey PubKeyEd25519) VerifyBytes(msg []byte, sig_ Signature) bool {
+	// make sure we use the same algorithm to sign
+	sig, ok := sig_.Unwrap().(SignatureEd25519)
+	if !ok {
+		return false
+	}
+	pubKeyBytes := [32]byte(pubKey)
+	sigBytes := [64]byte(sig)
+	return ed25519.Verify(&pubKeyBytes, msg, &sigBytes)
+}
+
+func (p PubKeyEd25519) MarshalJSON() ([]byte, error) {
+	return data.Encoder.Marshal(p[:])
+}
+
+func (p *PubKeyEd25519) UnmarshalJSON(enc []byte) error {
+	var ref []byte
+	err := data.Encoder.Unmarshal(&ref, enc)
+	copy(p[:], ref)
+	return err
+}
+
+// For use with golang/crypto/nacl/box
+// If error, returns nil.
+func (pubKey PubKeyEd25519) ToCurve25519() *[32]byte {
+	keyCurve25519, pubKeyBytes := new([32]byte), [32]byte(pubKey)
+	ok := extra25519.PublicKeyToCurve25519(keyCurve25519, &pubKeyBytes)
+	if !ok {
+		return nil
+	}
+	return keyCurve25519
+}
+
+func (pubKey PubKeyEd25519) String() string {
+	return cmn.Fmt("PubKeyEd25519{%X}", pubKey[:])
+}
+
+// Must return the full bytes in hex.
+// Used for map keying, etc.
+func (pubKey PubKeyEd25519) KeyString() string {
+	return cmn.Fmt("%X", pubKey[:])
+}
+
+func (pubKey PubKeyEd25519) Equals(other PubKey) bool {
+	if otherEd, ok := other.Unwrap().(PubKeyEd25519); ok {
+		return bytes.Equal(pubKey[:], otherEd[:])
+	} else {
+		return false
+	}
+}
+
+func (pubKey PubKeyEd25519) Wrap() PubKey {
+	return PubKey{pubKey}
+}
+
+//-------------------------------------
+
+var _ SignatureInner = SignatureEd25519{}
+
+// Implements Signature
+type SignatureEd25519 [64]byte
+
+func (sig SignatureEd25519) AssertIsSignatureInner() {}
+
+func (sig SignatureEd25519) Bytes() []byte {
+	return wire.BinaryBytes(Signature{sig})
+}
+
+func (sig SignatureEd25519) IsZero() bool { return len(sig) == 0 }
+
+func (sig SignatureEd25519) String() string { return fmt.Sprintf("/%X.../", cmn.Fingerprint(sig[:])) }
+
+func (sig SignatureEd25519) Equals(other Signature) bool {
+	if otherEd, ok := other.Unwrap().(SignatureEd25519); ok {
+		return bytes.Equal(sig[:], otherEd[:])
+	} else {
+		return false
+	}
+}
+
+func (sig SignatureEd25519) MarshalJSON() ([]byte, error) {
+	return data.Encoder.Marshal(sig[:])
+}
+
+func (sig *SignatureEd25519) UnmarshalJSON(enc []byte) error {
+	var ref []byte
+	err := data.Encoder.Unmarshal(&ref, enc)
+	copy(sig[:], ref)
+	return err
+}
+
+func (sig SignatureEd25519) Wrap() Signature {
+	return Signature{sig}
+}
